fix: resolve clippy warnings and mark story 7.4 done
- Fix collapsible if in config.rs
- Use char array patterns instead of closures
- Use RangeInclusive::contains for bounds checks
- Replace manual Default impls with #[derive(Default)]
- Use matches! macro for pattern matching
- Use Entry API for HashMap operations
- Remove unnecessary casts

Story 7.4 (saves_count metric) passes code review:
- All acceptance criteria verified
- Tests pass, clippy clean

Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-Claude)

+++ src/resume/backoff.rs
@@ -154,7 +154,7 @@ impl Backoff {
 
     fn base_delay_for_attempt(&self, attempt: u32) -> Duration {
         let attempt = attempt.max(1);
-        let exponent = attempt.saturating_sub(1).min(31) as u32;
+        let exponent = attempt.saturating_sub(1).min(31);
         let multiplier = 2u128.saturating_pow(exponent);
         let base_millis = self.config.base_delay.as_millis();
         let delay_millis = base_millis.saturating_mul(multiplier);
